requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.19
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
PyJWT==2.9.0
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
